requirements.txt
streamlit>=1.21.0
altair<5
pytest
google-generativeai>=0.7.0
